+++ CHANGELOG.md
@@ -4,7 +4,8 @@ xmchord Changelog
 V0.3.1 - Not released yet
 -------------------------
 * Add Claws Mail action: ◢ + F = Open "Search Folder" popup
-* Add action: ◢ + ' = Output name of application of focused window 
+* Add action: ◢ + ' = Output name of application of focused window
+* Add Gnome Terminal action: ◢ + Numpad = Insert 40 empty lines (helpful to separate sections of output) 
 
 V0.3.0 - 2020/12/20
 -------------------